Move one click hedging to separate file
